+++ dio_cache_interceptor/lib/src/dio_cache_interceptor_cache_utils.dart
@@ -124,6 +124,7 @@ extension _DioCacheInterceptorUtils on DioCacheInterceptor {
     return options.keyBuilder(
       url: request.uri,
       headers: request.getFlattenHeaders(),
+      body: request.data,
     );
   }
 }