+++ http_cache_core/lib/src/model/cache/cache_control.dart
@@ -110,6 +110,7 @@ class CacheControl {
           other.add(attribute);
         }
       }
+      scanner.scan(whitespace);
     }
 
     headerValues ??= [];