+++ http_cache_core/lib/src/model/cache/cache_options.dart
@@ -9,6 +9,7 @@ import 'cache_priority.dart';
 typedef CacheKeyBuilder = String Function({
   required Uri url,
   Map<String, String>? headers,
+  Object? body,
 });
 
 /// Options to apply to handle request and cache behaviour.
@@ -73,6 +74,7 @@ class CacheOptions {
   static String defaultCacheKeyBuilder({
     required Uri url,
     Map<String, String>? headers,
+    Object? body,
   }) {
     return _uuid.v5(Namespace.url.value, url.toString());
   }