+++ http_cache_isar_store/lib/src/store/cache_collection.g.dart
@@ -113,7 +113,7 @@ const CacheSchema = CollectionSchema(
   getId: _cacheGetId,
   getLinks: _cacheGetLinks,
   attach: _cacheAttach,
-  version: '3.1.0+1',
+  version: '3.3.0-dev.3',
 );
 
 int _cacheEstimateSize(