+++ http_cache_objectbox_store/lib/objectbox.g.dart
@@ -320,9 +320,8 @@ obx_int.ModelDefinition getObjectBoxModel() {
               const fb.Int64Reader().vTableGetNullable(buffer, rootOffset, 20);
           final keyParam = const fb.StringReader(asciiOptimization: true)
               .vTableGet(buffer, rootOffset, 6, '');
-          final contentParam =
-              const fb.ListReader<int>(fb.Int8Reader(), lazy: false)
-                  .vTableGetNullable(buffer, rootOffset, 8);
+          final contentParam = const fb.Uint8ListReader(lazy: false)
+              .vTableGetNullable(buffer, rootOffset, 8) as Uint8List?;
           final dateParam = dateValue == null
               ? null
               : DateTime.fromMillisecondsSinceEpoch(dateValue);
@@ -331,9 +330,8 @@ obx_int.ModelDefinition getObjectBoxModel() {
           final expiresParam = expiresValue == null
               ? null
               : DateTime.fromMillisecondsSinceEpoch(expiresValue);
-          final headersParam =
-              const fb.ListReader<int>(fb.Int8Reader(), lazy: false)
-                  .vTableGetNullable(buffer, rootOffset, 16);
+          final headersParam = const fb.Uint8ListReader(lazy: false)
+              .vTableGetNullable(buffer, rootOffset, 16) as Uint8List?;
           final lastModifiedParam =
               const fb.StringReader(asciiOptimization: true)
                   .vTableGetNullable(buffer, rootOffset, 18);