+++ dio_cache_interceptor/CHANGELOG.md
@@ -1,3 +1,12 @@
+## next
+- chore: Min SDK is now 3.6.0.
+
+## 4.0.5
+- chore: Updated dependencies.
+
+## 4.0.4
+- fix: Allow keyBuilder to depend on request body.
+
 ## 4.0.3
 - fix: BaseRequest now has `headers` getter instead of `headerValuesAsList`.
 - chore: Updated dependencies.