+++ http_cache_core/CHANGELOG.md
@@ -1,3 +1,12 @@
+## next
+- chore: Min SDK is now 3.6.0.
+
+## 1.1.3
+- fix: Allow optional whitespace in header values.
+
+## 1.1.2
+- fix: Allow keyBuilder to depend on request body.
+
 ## 1.1.1
 - fix: Send only one condition on request cache validation.
 - fix: BaseRequest now has `headers` getter instead of `headerValuesAsList`.